Re #73 Fix Baxter's base link to world
This commit adds a Fixed dummy link to Gazebo
which locks down Baxter, preventing him from
wandering over time with respect to Gazebo world.

I added this to the pedestal xacro to allow
for the future possibility of a mobile Baxter URDF
that would not need this fixed link (or pedistal).

Additionally, I consolidated the gazebo tags into
the baxter_base folder and removed the unused top
level gazebo xacro for baxter.