Re baxter_sim #66 by adding mass to null_gripper
When using the null gripper in Gazebo, the robot
would fail to load, instead showing just pieces
of Baxter on the ground. This was due to the
<side>_gripper_base and <side>_gripper links
having no mass at all, which I assume messes up
intertial calculations. By adding a nominal mass
of 0.0001, the [-nan -nan -nan] pose issue is
resolved.